Implement first lambda draft
This commit implements the first draft with working lambda
implementation.

LWS is now running on docker in order to run it safely as
root and to have a controlled environment to make it easier
to maintain.
For now Lambdas are executed on some kind of "hand crafted"
linux container managed by pkg/box.

Also get rid of viper. It does way more than what this
project needs and pollutes the dependencies which is making
builds a lot annoying.