Change gobox to signal when it is ready
We need a way to know when to start forwarding client requests
after launching a box instance.
To acheive that two different mechanisms where added to the gobox.

First it only proceeds when the lambda is ready, this means that
the rpc server run by the SDK is ready to accept requests. This is
done by trying to connect to the server a ew times.

The second mechanism is simple contract with the box system which
basically states that the box must write to the stdout or stderr
a predefined string ("#booted#").