Refactor llambda and box system
Massive commit to catch up with the last changes to
gobox and runbox as well as to add a box life cycle
handler to the llambda.

This should be splited into 100 commits... but at this
stage it is a bit difficult to do it because things are
quite unstable yet.