Make box 'thread safe'
The goal is to make box thread safe by applying the same system
already in place on LSQS, but this time, instead of creating new
types per each Action to pass to the channel, I'm passing to the
channel functions with the needed return values.
The reason I'm trying functions this time is to avoid the boiler
plate of creating new types every time a new action is created
and to avoid polluting the namespace. Lets see which one works
better.