Make box async
Unlike in LSQS, operation on box may take longer which
means that making action's processing sequencial will
hame impact in the number of request that it can serve.

To void this, all actions will be processed on a separate
goroutine and box map and instances map are now sync.Map.
Also instances are now 'instance' instead o '*instance'
to avoid unwanted changes.